[Refactor] #19 dto 디렉토리 패키징 전략 수정
