+++ src/main/java/com/donet/donet/donation/adapter/in/web/GetDonationListController.java
@@ -1,8 +1,8 @@
 package com.donet.donet.donation.adapter.in.web;
 
-import com.donet.donet.donation.application.port.in.response.GetFilteredDonationResponse;
+import com.donet.donet.donation.application.port.in.dto.response.GetFilteredDonationResponse;
 import com.donet.donet.donation.application.port.in.GetFilteredDonationUsecase;
-import com.donet.donet.donation.application.port.in.command.GetFilteredDonationCommand;
+import com.donet.donet.donation.application.port.in.dto.command.GetFilteredDonationCommand;
 import com.donet.donet.global.annotation.CurrentUserId;
 import com.donet.donet.global.response.BaseResponse;
 import lombok.RequiredArgsConstructor;
