+++ src/main/java/com/donet/donet/donation/application/port/in/GetDonationDetailUsecase.java
@@ -1,6 +1,6 @@
 package com.donet.donet.donation.application.port.in;
 
-import com.donet.donet.donation.application.port.in.response.GetDonationDetailResponse;
+import com.donet.donet.donation.application.port.in.dto.response.GetDonationDetailResponse;
 
 public interface GetDonationDetailUsecase {
     GetDonationDetailResponse getDonationDetail(Long donationId);
