+++ src/main/java/com/donet/donet/donation/application/port/in/GetFilteredDonationUsecase.java
@@ -1,7 +1,7 @@
 package com.donet.donet.donation.application.port.in;
 
-import com.donet.donet.donation.application.port.in.response.GetFilteredDonationResponse;
-import com.donet.donet.donation.application.port.in.command.GetFilteredDonationCommand;
+import com.donet.donet.donation.application.port.in.dto.response.GetFilteredDonationResponse;
+import com.donet.donet.donation.application.port.in.dto.command.GetFilteredDonationCommand;
 
 public interface GetFilteredDonationUsecase {
     GetFilteredDonationResponse getFilteredDonation(GetFilteredDonationCommand command);
